% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compile_model}
\alias{compile_model}
\title{compile_model}
\usage{
compile_model(id, compile_time = 300)
}
\arguments{
\item{id}{A string representing the id of the model.}

\item{compile_time}{An integer representing the amount of seconds we will wait before terminating the compilation process [300].}
}
\description{
Compile a model.
}
