% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_model}
\alias{create_model}
\title{create_model}
\usage{
create_model(name)
}
\arguments{
\item{name}{A string representing the name of the model.}
}
\description{
Function for creating a model. If successful returns the model's id number.
}
