% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{edit_model}
\alias{edit_model}
\title{edit_model}
\usage{
edit_model(id, code = "", data = NULL, name = "")
}
\arguments{
\item{id}{A string representing the id of the model.}

\item{code}{A string representing the code of the model or a filename that contains the model's code [""].}

\item{data}{A list containing the input data [NULL].}

\item{name}{A string representing the new name of the model [""].}
}
\description{
Edit a model.
}
