% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_details}
\alias{get_model_details}
\title{get_model_details}
\usage{
get_model_details(id)
}
\arguments{
\item{id}{A string representing the id of the model.}
}
\description{
Get details of a specific model.
}
