% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(id)
}
\arguments{
\item{id}{A string representing the id of the model.}
}
\description{
Function for retrieving results of sampling.
}
