% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_sampling_status}
\alias{get_sampling_status}
\title{get_sampling_status}
\usage{
get_sampling_status(id)
}
\arguments{
\item{id}{A string representing the id of the model.}
}
\description{
Function for checking the sampling status.
}
