% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{login}
\usage{
login(email, password = "")
}
\arguments{
\item{email}{A string representing an email of you cloudstan account.}

\item{password}{Password behind the providede email account. [""].}
}
\description{
Function for logging into the Cloudstan platform.
}
