% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_model}
\alias{sample_model}
\title{sample_model}
\usage{
sample_model(id, warmup = 1000, samples = 1000, async = FALSE)
}
\arguments{
\item{id}{A string representing the id of the model.}

\item{warmup}{An integer representing the amount of warmup steps [1000].}

\item{samples}{An integer representing the desired amount of samples [1000].}

\item{async}{A boolean for triggering asynchronous sampling [FALSE].}
}
\description{
Function for starting the sampling process.
}
