% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{login}
\usage{
login(email)
}
\arguments{
\item{email}{A string representing an email of you cloudstan account [""].}

\item{query}{A boolean whether the entered password will be remebered [FALSE].}
}
\description{
Function for logging into the Cloudstan platform.
}
